+++ src/RDF.ts
@@ -1,11 +1,7 @@
 import * as Persisted from './RDF/Persisted';
-import * as DocumentResource from './RDF/DocumentResource';
-import * as PersistedDocumentResource from './RDF/PersistedDocumentResource';
-import * as FragmentResource from './RDF/FragmentResource';
-import * as PersistedFragmentResource from './RDF/PersistedFragmentResource';
 import * as Literal from './RDF/Literal';
 import PropertyDescription from './RDF/PropertyDescription';
-import * as RDFDocument from './RDF/RDFDocument';
+import * as Document from './RDF/Document';
 import * as RDFNode from './RDF/RDFNode';
 import * as Resource from './RDF/Resource';
 import * as URI from './RDF/URI';
@@ -14,13 +10,9 @@ import * as Value from './RDF/Value';
 //@formatter:off
 export {
 	Persisted,
-	DocumentResource,
-	PersistedDocumentResource,
-	FragmentResource,
-	PersistedFragmentResource,
 	Literal,
 	PropertyDescription,
-	RDFDocument as Document,
+	Document,
 	RDFNode as Node,
 	Resource,
 	URI,